refactor(ppu): split mod.rs into separate modules for better organization
Split the monolithic PPU module (2334 lines) into focused submodules:

- constants.rs (28 lines): PPU constants (screen size, tile size, etc.)
- memory.rs (184 lines): Memory access (VRAM, palette, nametables)
- registers.rs (198 lines): Register read/write handling
- rendering.rs (205 lines): Background rendering pipeline
- mod.rs (1749 lines): PPU struct, public API, and tests

Benefits:
- Improved readability with clear separation of concerns
- Easier maintenance and navigation
- Better support for parallel development
- Simplified future additions (e.g., sprite rendering)

All 70 tests pass. No functional changes.
